requirements.txt
monai==1.5.0
nibabel==5.3.2
numpy==2.1.2
pandas==2.3.0
prettytable==3.16.0
scikit-learn==1.7.0
scipy==1.15.3
SimpleITK==2.5.2
surface-distance==0.1
timm==1.0.16
torchmetrics==1.7.3